fix(arcium): rename close_position_v3, drop u128/i128 casts, widen u16 to u64
Circuit changes (per Cryptanalyst guidance — narrowing casts in Arcis trigger
AbortedComputation on the cluster):

- close_position: removed u128/i128 intermediates from PnL and fee calcs,
  restructured to stay in i64/u64 throughout via pre-division (matches the
  pattern already in liquidation_check). Renamed close_position_v2 to
  close_position_v3 because the on-chain v2 comp-def is immutable after the
  trading_fee_bps signature change (u16 to u64).
- settle_private_position: same cast removal, widened trading_fee_bps.
- liquidation_check: widened liquidation_threshold_bps u16 to u64.

Program updates:
- All v3 references aligned across lib.rs, handlers, callbacks, init_comp_defs,
  tpsl. Output types renamed to ClosePositionV3Output / ClosePositionV3Callback.
- Handlers widen plaintext params via u64::from() before the ArgBuilder so
  circuit signatures stay on the safe integer width.
- check_liquidation handler updated to send u64 threshold.

Scripts: 7 files updated for the v3 rename (init-comp-defs, init-all-markets,
sync-market-comp-defs, upload-circuits, smoke-test-devnet, check-cluster-health,
wsl-arcium-build alias).

UI:
- TradeConfirmationModal.isRetryableError fixed: MPC aborts now correctly
  classified as retryable. Program callbacks already leave positions in a
  clean retryable state (open: Closed with no margin lock; close: back to Open).
- BottomPositionsPanel: Funding History tab fully wired with table layout,
  Balances tab redesign with inline Wallet/Trading Account sections,
  Deposit/Withdraw buttons launching CollateralModal on the right tab,
  retry toast on close errors that are retryable.

IDL regenerated by build, picks up V3 names.

Open follow-up: check_liquidation_v2 also has the same on-chain immutability
issue after the threshold widening — needs a parallel rename to
check_liquidation_v3 + comp-def re-init. Tracked separately to keep this
commit scoped to the close path which is the user-visible blocker.

+++ encrypted-ixs/src/liquidation_check.rs
@@ -20,7 +20,7 @@ mod liquidation_check_circuit {
     pub fn check_liquidation_v2(
         position: Enc<Shared, (u64, u64, u8, u8, u64)>,
         mark_price: u64,
-        liquidation_threshold_bps: u16,
+        liquidation_threshold_bps: u64,
     ) -> (bool, u64, u64) {
         let pos = position.to_arcis();
 
@@ -50,7 +50,7 @@ mod liquidation_check_circuit {
         // notional = size * mark_price / 10^6 â€” same pre-division pattern.
         let notional = (size / 1000) * (mark_price / 1000);
         // maintenance = notional * threshold_bps / 10000 â€” exact division, no rebasing.
-        let maintenance = notional * liquidation_threshold_bps as u64 / 10000;
+        let maintenance = notional * liquidation_threshold_bps / 10000;
 
         let should_liquidate = equity < maintenance;
         let revealed_margin = if should_liquidate { margin } else { 0 };
